refactor(workorder,general): Refatorando package de work order e outros ajustes de padronização

+++ src/main/java/com/mecanicadm/mecanicadm_api/infra/features/workorder/persistence/jpa/WorkOrderLaborItemRepositoryImpl.java
@@ -1,12 +1,16 @@
 package com.mecanicadm.mecanicadm_api.infra.features.workorder.persistence.jpa;
 
-import com.mecanicadm.mecanicadm_api.core.workorders.domain.WorkOrderLaborItem;
-import com.mecanicadm.mecanicadm_api.core.workorders.domain.port.WorkOrderLaborItemGateway;
-import com.mecanicadm.mecanicadm_api.infra.features.workorder.persistence.jpa.mapper.WorkOrderLaborItemJpaMapper;
+import com.mecanicadm.mecanicadm_api.core.workorder.domain.WorkOrderLaborItem;
+import com.mecanicadm.mecanicadm_api.core.workorder.domain.port.WorkOrderLaborItemGateway;
+import com.mecanicadm.mecanicadm_api.shared.exception.TechnicalException;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 import java.util.UUID;
 
+import static java.util.Objects.isNull;
+
+@Repository
 public class WorkOrderLaborItemRepositoryImpl implements WorkOrderLaborItemGateway {
 
     private final WorkOrderLaborItemJpaRepository jpaRepository;
@@ -22,6 +26,9 @@ public class WorkOrderLaborItemRepositoryImpl implements WorkOrderLaborItemGatew
 
     @Override
     public void delete(WorkOrderLaborItem laborItem) {
+        if (isNull(laborItem)) {
+            throw new TechnicalException("error.technical.entity.null", "WorkOrderLaborItem", "exclusão");
+        }
         jpaRepository.delete(WorkOrderLaborItemJpaMapper.toEntity(laborItem));
     }
 }
